Split up tests into unit test suite and integration test suite
