+++ phpunit.xml.dist
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
-
 <phpunit backupGlobals="false"
          backupStaticAttributes="false"
          colors="true"
@@ -12,8 +11,11 @@
         <log type="coverage-clover" target="build/logs/clover.xml"/>
     </logging>
     <testsuites>
-        <testsuite name="Test Suite">
-            <directory>test/</directory>
+        <testsuite name="Unit Test Suite">
+            <directory>tests/unit</directory>
+        </testsuite>
+        <testsuite name="Integration Test Suite">
+            <directory>tests/integration</directory>
         </testsuite>
     </testsuites>
 </phpunit>
